+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
-- /
+
+- fixed log spam when a custom ritual renderer is erroring
 
 ## [3.7.1] - 2026-04-09